requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
stripe==9.12.0
python-dotenv==1.0.1
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.0.1
